Organise project file system
