+++ src/components/styles/App.module.css
@@ -5,6 +5,10 @@
 .overlay {
   background-color: rgba(255, 255, 255, 0);
   position: absolute;
+  bottom: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 4;
 }
 
 .recipe-card:hover .overlay{
@@ -31,4 +35,5 @@
   font-size: 2rem;
   font-weight: 100;
   color: rgba(0, 0, 0, 0.9);
+  background: transparent;
 }
