+++ src/store/store.js
@@ -1,6 +1,6 @@
 import { createStore } from "redux";
-import reducer from "./reducer.js";
-import getMeals from "./helpers/getMeals.js";
+import reducer from "../reducers/reducer";
+import getMeals from "../helpers/getMeals.js";
 
 export default (async function() {
   let meals = await getMeals().then(meals =>
